Readme, composer update, renamed files, gitignore.
