+++ .gitignore
@@ -1,3 +1,5 @@
+vendor/*
+
 #################
 ## Eclipse
 #################
