+++ src/Util/Stack.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace JonLawrence\Util;
+namespace EquationOS\Util;
 
 /**
  * Basic Stack Class.
@@ -15,7 +15,7 @@ namespace JonLawrence\Util;
  * @subpackage EOS
  * @version 2.0
  */
-class phpStack {
+class Stack {
 	private $index;
 	private $locArray;
 
@@ -58,7 +58,7 @@ class phpStack {
 	/**
 	 * Push
 	 *
-	 * Alias of {@see phpStack::poke()}
+	 * Alias of {@see Stack::poke()}
 	 * Adds element to the stack
 	 *
 	 * @param Mixed Element to add
